+++ packages/cache-handler/src/handlers/server.ts
@@ -1,5 +1,3 @@
-import { replaceJsonWithBase64, reviveFromBase64Representation } from '@neshca/json-replacer-reviver';
-
 import type { CacheHandlerValue, Handler } from '../cache-handler';
 
 export type ServerCacheHandlerOptions = {
@@ -26,7 +24,7 @@ export type ServerCacheHandlerOptions = {
  *
  * @example
  * ```js
- * const serverCache = createCache({
+ * const serverHandler = createHandler({
  *   baseUrl: 'http://localhost:8080/',
  * });
  * ```
@@ -36,7 +34,7 @@ export type ServerCacheHandlerOptions = {
  * - the `set` method allows setting a value in the server cache.
  * - the `revalidateTag` methods are used for handling tag-based cache revalidation.
  */
-export default function createCache({ baseUrl, timeoutMs }: ServerCacheHandlerOptions): Handler {
+export default function createHandler({ baseUrl, timeoutMs }: ServerCacheHandlerOptions): Handler {
     return {
         name: 'server',
         async get(key) {
@@ -62,19 +60,16 @@ export default function createCache({ baseUrl, timeoutMs }: ServerCacheHandlerOp
 
             const string = await response.text();
 
-            return JSON.parse(string, reviveFromBase64Representation) as CacheHandlerValue;
+            return JSON.parse(string) as CacheHandlerValue;
         },
         async set(key, cacheHandlerValue) {
             const url = new URL('/set', baseUrl);
 
-            const isRouteKind = cacheHandlerValue.value?.kind === 'ROUTE';
+            const _isRouteKind = cacheHandlerValue.value?.kind === 'ROUTE';
 
             const response = await fetch(url, {
                 method: 'POST',
-                body: JSON.stringify([
-                    key,
-                    JSON.stringify(cacheHandlerValue, isRouteKind ? replaceJsonWithBase64 : undefined),
-                ]),
+                body: JSON.stringify([key, JSON.stringify(cacheHandlerValue)]),
                 headers: {
                     'Content-Type': 'application/json',
                 },